move func, explicity true flag
