+++ skyrl-train/skyrl_train/generators/utils.py
@@ -1,8 +1,8 @@
 import torch
-from typing import List, Tuple, Union, Optional
+from typing import List, Tuple, Union, Optional, Dict, Any
 from collections import defaultdict
 import numpy as np
-from skyrl_train.generators.base import GeneratorOutput
+from skyrl_train.generators.base import GeneratorOutput, GeneratorInput, TrajectoryID, BatchMetadata, TrainingPhase
 
 CUSTOM_CHAT_TEMPLATES = {
     # chat template for qwen3 thinking mode to remove think tokens similar to generation phase
@@ -152,3 +152,63 @@ def get_rollout_metrics(responses: List[List[int]], rewards: Union[List[float], 
         "generate/avg_tokens_non_zero_rewards": avg_tokens_non_zero_rewards.item(),
         "generate/avg_tokens_zero_rewards": avg_tokens_zero_rewards.item(),
     }
+
+
+def prepare_generator_input(
+    n_samples_per_prompt: int,
+    prompts: List[Any],
+    sampling_params: Dict[str, Any],
+    default_env_class: str,
+    training_phase: TrainingPhase,
+    global_step: int,
+) -> Tuple[GeneratorInput, List[str]]:
+    """Prepares the generator input for training and eval
+
+    Args:
+        n_samples_per_prompt (int): how many samples to create per prompt
+        prompts (List[Any]): list of prompts
+        sampling_params (Dict[str, Any]): sampling parameters
+        default_env_class (str): env class to use if env class missing from prompts
+        training_phase (TrainingPhase): training or eval
+        global_step (int): current global step
+
+    Returns:
+        Tuple[GeneratorInput, List[str]]: generator input and list of uuids
+    """
+
+    all_prompts = sum([[prompt["prompt"]] * n_samples_per_prompt for prompt in prompts], [])
+    all_envs = sum(
+        [
+            [prompt["env_class"] if prompt["env_class"] is not None else default_env_class] * n_samples_per_prompt
+            for prompt in prompts
+        ],
+        [],
+    )
+
+    # all the other columns are env_extras
+    env_extras = sum(
+        [[prompt["env_extras"]] * n_samples_per_prompt for prompt in prompts],
+        [],
+    )
+
+    # Create TrajectoryID objects - one UID per row, repetition_id for multiple samples
+    trajectory_ids = []
+    uids = []
+    for _, prompt in enumerate(prompts):
+        uid: str = prompt["uid"]
+
+        # Create TrajectoryID for each repetition
+        for repetition_id in range(n_samples_per_prompt):
+            trajectory_ids.append(TrajectoryID(instance_id=uid, repetition_id=repetition_id))
+            uids.append(uid)
+
+    generator_input: GeneratorInput = {
+        "prompts": all_prompts,
+        "env_classes": all_envs,
+        "env_extras": env_extras,
+        "sampling_params": sampling_params,
+        "trajectory_ids": trajectory_ids,
+        "batch_metadata": BatchMetadata(global_step=global_step, training_phase=training_phase),
+    }
+
+    return generator_input, uids
